chore: Refine distribution
