+++ built-in.js
@@ -1,0 +1,5 @@
+import EnhancedTextAreaBuiltIn, { defineElement } from './dist/built-in/index.js';
+
+export default EnhancedTextAreaBuiltIn;
+
+export { defineElement };
