+++ index.js
@@ -1,0 +1,5 @@
+import EnhancedTextAreaBuiltIn from './dist/built-in/index.js';
+
+export {
+  EnhancedTextAreaBuiltIn
+};
