+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@leofavre/enhanced-textarea",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "description": "Enhanced Text Area",
-  "main": "dist/index.js",
+  "main": "index.js",
   "scripts": {
     "build": "snowpack build",
     "cypress": "cypress run",
