+++ public/index.html
@@ -6,7 +6,7 @@
   <title>Enhanced Text Area</title>
   <link rel="stylesheet" href="./index.css">
   <script type="module">
-    import { defineElement } from './built-in.js';
+    import { defineElement } from './built-in/index.js';
     window.defineElement = defineElement;
     const { port = '8080' } = new URL(document.URL);
     if (Number(port) > 8000) defineElement();
