--- src/built-in.js
@@ -1,11 +1,0 @@
-import EnhancedTextAreaBuiltIn from './built-in/index.js';
-
-export const defineElement = (name = 'enhanced-textarea') => {
-  customElements.define(
-    name,
-    EnhancedTextAreaBuiltIn,
-    { extends: 'textarea' }
-  );
-};
-
-export default EnhancedTextAreaBuiltIn;
