+++ src/built-in/index.js
@@ -1,5 +1,13 @@
 import EnhancedTextAreaBuiltInFactory from './factory.js';
 
+export const defineElement = (name = 'enhanced-textarea') => {
+  customElements.define(
+    name,
+    EnhancedTextAreaBuiltIn,
+    { extends: 'textarea' }
+  );
+};
+
 const EnhancedTextAreaBuiltIn =
   EnhancedTextAreaBuiltInFactory(HTMLTextAreaElement);
