--- src/index.js
@@ -1,5 +1,0 @@
-import EnhancedTextAreaBuiltIn from './built-in/index.js';
-
-export {
-  EnhancedTextAreaBuiltIn
-};
